[CALCITE-2483] Druid adapter, when querying Druid segment metadata, throws when row number is larger than Integer.MAX_VALUE

+++ druid/src/main/java/org/apache/calcite/adapter/druid/DruidConnectionImpl.java
@@ -719,8 +719,8 @@ class DruidConnectionImpl implements DruidConnection {
     public String id;
     public List<String> intervals;
     public Map<String, JsonColumn> columns;
-    public int size;
-    public int numRows;
+    public long size;
+    public long numRows;
     public Map<String, JsonAggregator> aggregators;
   }
